Move models to common Handle upn check for JUSTIN users
