+++ backend/jumwebapi/Features/Participants/Controllers/ParticipantController.cs
@@ -1,12 +1,11 @@
 namespace jumwebapi.Features.Participants.Controllers;
 
-using jumwebapi.Features.Participants.Models;
-using Microsoft.AspNetCore.Mvc;
+using System.Text.Json;
+using jumwebapi.Features.Participants.Queries;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
-using jumwebapi.Features.Participants.Queries;
+using Microsoft.AspNetCore.Mvc;
 using Serilog;
-using System.Text.Json;
 
 
 
@@ -27,11 +26,11 @@ public class ParticipantController : ControllerBase
     [ProducesResponseType(StatusCodes.Status200OK)]
     [ProducesResponseType(StatusCodes.Status400BadRequest)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
-    public async Task<IActionResult> ParticipantByUsername([Required]string username)
+    public async Task<IActionResult> ParticipantByUsername([Required] string username)
     {
         Log.Logger.Information("Getting participant info for {0}", username);
         var participant = await this.mediator.Send(new GetParticipantByUsernameQuery(username));
-        Log.Logger.Information("Got {0}", JsonSerializer.Serialize(participant));
+        Log.Logger.Information("Got participant {0}", JsonSerializer.Serialize(participant));
 
         return new JsonResult(participant);
     }
