+++ backend/jumwebapi/Infrastructure/HttpClients/JustinParticipantClient/JustinParticipantClient.cs
@@ -1,4 +1,6 @@
 namespace jumwebapi.Infrastructure.HttpClients.JustinParticipant;
+
+using global::Common.Models.JUSTIN;
 using jumwebapi.Features.Participants.Models;
 
 public class JustinParticipantClient : BaseClient, IJustinParticipantClient
