+++ backend/webapi/Features/AccessRequests/DigitalEvidenceUpdate.cs
@@ -3,6 +3,7 @@ namespace Pidp.Features.AccessRequests;
 using System;
 using System.Diagnostics;
 using Common.Models;
+using Common.Models.JUSTIN;
 using Common.Models.Notification;
 using DomainResults.Common;
 using FluentValidation;
