+++ backend/webapi/Features/Parties/ProfileStatus.cs
@@ -9,6 +9,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using common.Constants.Auth;
 using Common.Models.EDT;
+using Common.Models.JUSTIN;
 using FluentValidation;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.EntityFrameworkCore;
